Fix CnblogProcess to fit Export multiFormat document. Add EpubProcess.cs to generate Epub,not finished now.

+++ Blog.Process/Interfaces/DownloadStringTaskAsyncExProgress.cs
@@ -2,7 +2,7 @@
 {
     public class DownloadStringTaskAsyncExProgress
     {
-        public int ProgressPercentage { get; set; }
+        public float ProgressPercentage { get; set; }
 
         public string Text { get; set; }
     }
